Update TS migration to current dev, move to a common package, rename to frontend
