--- core/search-component/src/main/components/app/index.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import {WithFuzzySearchFilter} from '../search/search';
-import './index.scss';
-
-const App: React.FC = () => (
-    <div className="search-content">
-        <WithFuzzySearchFilter/>
-    </div>
-)
-
-export default App
