--- core/search-component/types/index.d.ts
@@ -1,1 +1,0 @@
-/// <reference types="./@jetbrains" />
