+++ plugins/base/frontend/.gitignore
