+++ plugins/base/frontend/src/main/components/search/types.ts
