+++ plugins/base/frontend/src/main/scss/index.scss
