+++ plugins/base/frontend/src/main/types/@jetbrains/index.d.ts
