+++ plugins/base/frontend/stylelint.config.js
