+++ plugins/base/frontend/webpack.config.js
